clean up of rake test runs. Closes GH-231
