+++ Jasper.sln
@@ -67,6 +67,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pinger", "src\Pinger\Pinger
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ponger", "src\Ponger\Ponger.csproj", "{7BF5E459-D6B6-45C5-8EFB-6CCCB9467ADC}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Jasper.LightningDb.Testing", "src\Jasper.LightningDb.Testing\Jasper.LightningDb.Testing.csproj", "{92EEB43A-C8E7-4E15-9C5B-B935ED304DAE}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "LightningDb", "LightningDb", "{6E13AC22-9BB2-4164-AB14-C5087F07C8E1}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "LightningDb", "LightningDb", "{3BF9DD90-E67F-4CEC-82DB-C5E7338962B8}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -389,6 +395,18 @@ Global
 		{7BF5E459-D6B6-45C5-8EFB-6CCCB9467ADC}.Release|x64.Build.0 = Release|x64
 		{7BF5E459-D6B6-45C5-8EFB-6CCCB9467ADC}.Release|x86.ActiveCfg = Release|x86
 		{7BF5E459-D6B6-45C5-8EFB-6CCCB9467ADC}.Release|x86.Build.0 = Release|x86
+		{92EEB43A-C8E7-4E15-9C5B-B935ED304DAE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{92EEB43A-C8E7-4E15-9C5B-B935ED304DAE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{92EEB43A-C8E7-4E15-9C5B-B935ED304DAE}.Debug|x64.ActiveCfg = Debug|x64
+		{92EEB43A-C8E7-4E15-9C5B-B935ED304DAE}.Debug|x64.Build.0 = Debug|x64
+		{92EEB43A-C8E7-4E15-9C5B-B935ED304DAE}.Debug|x86.ActiveCfg = Debug|x86
+		{92EEB43A-C8E7-4E15-9C5B-B935ED304DAE}.Debug|x86.Build.0 = Debug|x86
+		{92EEB43A-C8E7-4E15-9C5B-B935ED304DAE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{92EEB43A-C8E7-4E15-9C5B-B935ED304DAE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{92EEB43A-C8E7-4E15-9C5B-B935ED304DAE}.Release|x64.ActiveCfg = Release|x64
+		{92EEB43A-C8E7-4E15-9C5B-B935ED304DAE}.Release|x64.Build.0 = Release|x64
+		{92EEB43A-C8E7-4E15-9C5B-B935ED304DAE}.Release|x86.ActiveCfg = Release|x86
+		{92EEB43A-C8E7-4E15-9C5B-B935ED304DAE}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -412,6 +430,9 @@ Global
 		{80F9678C-6935-4048-AACF-41642A48282A} = {A806095A-9D66-4D55-8662-1FC67E90F6FB}
 		{A52C1084-C382-46DD-8978-02336EF16B1A} = {A806095A-9D66-4D55-8662-1FC67E90F6FB}
 		{7BF5E459-D6B6-45C5-8EFB-6CCCB9467ADC} = {A806095A-9D66-4D55-8662-1FC67E90F6FB}
+		{6E13AC22-9BB2-4164-AB14-C5087F07C8E1} = {2A0AC53D-ECE1-43E5-8FB5-497A03E70641}
+		{8852E247-D855-4018-8D6E-CED628CDBFF7} = {3BF9DD90-E67F-4CEC-82DB-C5E7338962B8}
+		{92EEB43A-C8E7-4E15-9C5B-B935ED304DAE} = {3BF9DD90-E67F-4CEC-82DB-C5E7338962B8}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {D92D723F-44EC-4C1E-AAC3-C162FCEAAA08}
