+++ src/IntegrationTests/AspNetCoreIntegration/composing_request_delegate_order_within_aspnetcore_bootstrapping.cs
@@ -1,38 +1,47 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Alba;
-using AlbaForJasper;
+using Jasper;
 using Jasper.Http;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Xunit;
 
-namespace Jasper.Testing.AspNetCoreIntegration
+namespace IntegrationTests.AspNetCoreIntegration
 {
-    public class composing_request_delegate_order : IDisposable
+    public class composing_request_delegate_order_within_aspnetcore_bootstrapping : IDisposable
     {
-        private readonly JasperRegistry theRegistry = new JasperRegistry();
+        private SystemUnderTest _alba;
 
-        private readonly Lazy<JasperRuntime> _runtime;
-
-        public composing_request_delegate_order()
+        private void theAppIs(Action<IApplicationBuilder> configure)
         {
-            theRegistry.Handlers.DisableConventionalDiscovery(true);
-            _runtime = new Lazy<JasperRuntime>(() => JasperRuntime.For(theRegistry));
+            var registry = new JasperRegistry();
+            registry.Handlers.DisableConventionalDiscovery(true);
+
+            _alba = SystemUnderTest.For(builder =>
+            {
+                builder
+                    .UseServer(new NulloServer())
+                    .Configure(configure)
+                    .UseJasper(registry);
+            });
+
         }
 
         public void Dispose()
         {
-            if (_runtime.IsValueCreated)
-            {
-                _runtime.Value.Dispose();
-            }
+            _alba?.Dispose();
         }
 
         private Task<IScenarioResult> scenario(Action<Scenario> configure)
         {
-            return _runtime.Value.Scenario(configure);
+            if (_alba == null)
+            {
+                theAppIs(_ => {});
+            }
+
+            return _alba.Scenario(configure);
         }
 
         [Fact]
@@ -60,7 +69,7 @@ namespace Jasper.Testing.AspNetCoreIntegration
         [Fact]
         public Task use_middleware_in_front()
         {
-            theRegistry.Http.Configure(app =>
+            theAppIs(app =>
             {
                 app.Use(next =>
                 {
@@ -85,7 +94,7 @@ namespace Jasper.Testing.AspNetCoreIntegration
         [Fact]
         public Task default_404_behavior_with_middleware_in_front()
         {
-            theRegistry.Http.Configure(app =>
+            theAppIs(app =>
             {
                 app.Use(next =>
                 {
@@ -97,6 +106,7 @@ namespace Jasper.Testing.AspNetCoreIntegration
                     };
                 });
             });
+
 
             return scenario(_ =>
             {
@@ -111,7 +121,7 @@ namespace Jasper.Testing.AspNetCoreIntegration
         [Fact]
         public async Task use_middleware_behind_jasper()
         {
-            theRegistry.Http.Configure(app =>
+            theAppIs(app =>
             {
                 app.AddJasper();
 
@@ -140,7 +150,7 @@ namespace Jasper.Testing.AspNetCoreIntegration
         [Fact]
         public async Task put_jasper_in_the_middle()
         {
-            theRegistry.Http.Configure(app =>
+            theAppIs(app =>
             {
                 app.Use(next =>
                 {
@@ -180,16 +190,6 @@ namespace Jasper.Testing.AspNetCoreIntegration
             });
 
 
-        }
-    }
-
-    public class TracingEndpoint
-    {
-        public static Task get_jasper_trace(HttpContext context)
-        {
-            context.Response.ContentType = "text/plain";
-            context.Response.StatusCode = 200;
-            return context.Response.WriteAsync("jasper was called");
         }
     }
 }
