+++ src/Jasper.Diagnostics/resources/manifest.json
@@ -1,4 +1,4 @@
 {
-  "bundle.css": "css/bundle.2e64cfe751d0fb78669e.css",
-  "bundle.js": "js/bundle.2e64cfe751d0fb78669e.js"
+  "bundle.css": "css/bundle.5e2fcc2e8d5a57cfdab9.css",
+  "bundle.js": "js/bundle.5e2fcc2e8d5a57cfdab9.js"
 }
