+++ src/Jasper.LightningDb.Testing/Jasper.LightningDb.Testing.csproj
@@ -3,14 +3,12 @@
     <TargetFramework>netcoreapp1.1</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Server.Kestrel" Version="1.1.3" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.3.0" />
     <PackageReference Include="Shouldly" Version="2.8.3" />
-    <PackageReference Include="System.Net.WebSockets.Client" Version="4.3.1" />
     <PackageReference Include="xunit" Version="2.3.0" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.3.0" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Jasper.WebSockets\Jasper.WebSockets.csproj" />
+    <ProjectReference Include="..\Jasper.LightningDb\Jasper.LightningDb.csproj" />
   </ItemGroup>
 </Project>
